adjust governance text (#91)
There is never an Admin meeting, so there is no Admin moderator or
agenda items. Remove that text as irrelevant. It was probably
copy/pasted from another doc.

The paragraph on votes seems to waver
between requiring a majority vote of all members vs. a majority vote on
both TSC and CommComm. Go with requiring the vote to pass on both
committees, as that is more stringent. (If it passes on both committees
with a majority, then there is also a majority overall. The reverse is
not necessarily true.) Next, remove the text about "simple majority" and
whatnot. CommComm and TSC should use whatever decision-making rules
around voting that they each have. That *may* be simple majority, but if
it is not, that's OK. It's also fine if CommComm and TSC have different
rules around votes. They should apply their own rules to their own
votes.