update process for travel requests (#105)
We currently have somewhat of a problem with travel requests in that
there is no clear way to know when one is accepted. This was not a
problem when it was entirely up to the TSC, but now that this is jointly
administered by TSC and CommComm, many of the travel request PRs appear
stalled.

Here is my hopefully-uncontroversial proposal for unstalling. It would
be great to have pile-on LGTMs from both TSC and CommComm on this so we
can land it relatively quickly and get some resolution on the pending
travel requests.