Update travel fund request for @ChALkeR
This increases the requested amount from $1500 to $1660,
mostly due to unaccaunted (at first) lodging taxes and
visa costs being larger then anticipated.

Travel: ~$910
Visa costs: ~$210
Lodging: ~$540

This is not the final amount yet, that would be available after
the event is over, but the final amount is expected to be lower
than this (most likely approximately equal, as most of this
is finalized).

This is a ~$10 increase for travel, ~$60 increase for visa costs,
and ~$90 increase for lodging costs over the previosly approved
amount.

Refs: https://github.com/nodejs/admin/pull/165