Update MEMBER_TRAVEL_FUND.md
Slight further change to the file name. I learned that people will make multiple requests in a year so I added a requirement to put the event name in the expense file. This will be helpful for the accounts team to differentiate between seemingly similar requests.

Clarified requesters need to sign the sheet before submitting.