Update the logistics of filing expenses
As mentioned in issue #379, I'd like to propose some updates to the
reimbursement process to make it faster and less error-prone.  This was tested
in a limited way during the last cycle, and it worked well.

At a high level, this provides a new option for submitting expenses that doesn't
require spreadsheets or a zip file of receipts (although it's still available if
someone really wants to use it).  Rather, it encourages people to use a free
Expensify account to organize and submit their receipts.

From the travel fund recipient's perspective, the biggest advantage of this
approach is that they can use Expensify's phone app to capture receipts as
expenses are incurred during the trip.  At the end of the travel, they
simply log into Expensify, create an expense report, and send it to the
operations@openjsf.org alias for payment.  This will also enable the Foundation
to process payments faster, because the submissions will be uniform.

The old process would still be available as a backup, of course.  However, it's
cumbersome and more time consuming, and will take longer to process.

Signed-off-by: Brian Warner <brian@bdwarner.com>