Update the GItHub management policy for Moderation
We now have access to a beta feature that permits a moderation role in the org. (Thanks, @bnb!) It doesn't quite allow the moderation team to do *everything* they currently can do, but it seems close and I'd like to give it a try. We can always flip things back if it doesn't work out.

@nodejs/tsc @nodejs/moderation