Update MemberExpectations.md
The fact that people are supposed to treat each other respectfully is a given and is in our code of conduct ("Being respectful of differing opinions, viewpoints, and experiences"). Repeating it here is unnecessary, makes this document longer than it needs to be, and increases the possibility of a conflict between this document and the code of conduct.