chore: remove "self-moderate" from Moderation Policy
The term "self-moderate" is probably clear in the context of the
Moderation Policy, but it has become a bit of our project's jargon and
it's not clear that people generally understand what the term means. It
is unnecessary in the Moderation Policy context. I think "edit their
comment" is more clear in the context used here.